service/requirements.txt
apscheduler
Flask
mysql-connector-python
requests
requests_toolbelt
